+++ src/bagels/components/fields.py
@@ -5,6 +5,7 @@ from textual.widgets import Input, Label, Static, Switch
 
 from bagels.components.autocomplete import AutoComplete, Dropdown, DropdownItem
 from bagels.forms.form import Form, FormField
+from bagels.managers.categories import get_category_by_id
 from bagels.utils.format import parse_formula_expression
 
 _RESTRICT_TYPES = {
@@ -40,6 +41,12 @@ class Field(Static):
 
         # Create base input widget
         self.input = Input(placeholder=field.placeholder or "", id=f"field-{field.key}")
+        self.autocomplete_postfix_display_label = Label(
+            "",
+            classes="autocomplete-postfix-display-label",
+            id="autocomplete-postfix-display-label",
+        )
+        self.id = f"field-{field.key}-controller"
 
         # Configure input based on field type
         match self.field.type:
@@ -50,24 +57,42 @@ class Field(Static):
                 self.input.value = field.default_value or ""
 
             case "autocomplete":
-                self.input.heldValue = field.default_value or ""
-                self.input.value = field.default_value_text or field.default_value or ""
+                default_value = field.default_value or ""
+                self.input.heldValue = default_value
+                self.input.value = field.default_value_text or default_value or ""
+                for index, option in enumerate(field.options.items):
+                    if option.value == default_value:
+                        self.handle_select_index(index)
+                        break
 
             case type_ if type_ != "boolean":
                 self.input.value = field.default_value or ""
 
+    def handle_select_index(self, index: int) -> None:
+        """Handler for (externally) selecting an autocomplete option"""
+        # Find matching option and set held value
+        if index == -1:
+            return
+        # Get the selected item directly using the index
+        selected_item = self.field.options.items[index]
+        self.input.heldValue = selected_item.value
+
+        if self.field.key != "categoryId":
+            self.autocomplete_postfix_display_label.update("")
+            return
+
+        # Update postfix display for category fields
+        postfix_display = f" {selected_item.postfix}"
+        category = get_category_by_id(selected_item.value)
+        if not category:
+            return
+        self.autocomplete_postfix_display_label.update(
+            f"[{category.parentCategory.color}]{postfix_display}[/]"
+        )
+
     def on_auto_complete_selected(self, event: AutoComplete.Selected) -> None:
         """Handle autocomplete selection"""
-        # self.screen.focus_next()
-
-        # Find matching option and set held value
-        if event.index == -1:
-            return
-        for i, item in enumerate(self.field.options.items):
-            selected_dropdown_index = event.index
-            if i == selected_dropdown_index:
-                self.input.heldValue = item.value
-                break
+        self.handle_select_index(event.index)
 
     def on_input_changed(self, event: Input.Changed):
         if self.field.type == "number":
@@ -110,13 +135,15 @@ class Field(Static):
                     show_when_empty=self.field.autocomplete_selector,
                 )
 
-                yield AutoComplete(
-                    self.input,
-                    dropdown,
-                    classes="field-autocomplete",
-                    create_action=self.field.create_action,
-                    backspace_clears=self.field.autocomplete_selector,
-                )
+                with Container(classes="autocomplete-container"):
+                    yield AutoComplete(
+                        self.input,
+                        dropdown,
+                        classes="field-autocomplete",
+                        create_action=self.field.create_action,
+                        backspace_clears=self.field.autocomplete_selector,
+                    )
+                    yield self.autocomplete_postfix_display_label
 
             elif self.field.type == "boolean":
                 with Container(classes="switch-group"):